// src/app/interfaces/customer-edit-dto.ts
export interface CustomerEditDto {
    id?: number|undefined;
    givenName?: string|undefined;
    surname?: string|undefined;
    address?: string|undefined;
    country?: string|undefined;
    phone?: string|undefined;
    email?: string|undefined;
    passportNo?: string|undefined;
    userId?: number|undefined;
    customerTypeId?: number|undefined;

}